move getFloat/Bool to yamlHelper
- and make yamlHelper includable in more than one file
